Airboss: SetMPWireCorrection to 15 for Supercarriers
This commit sets the wire correction value to 15.

Background:

Prior to this update, airboss wass off regarding which wire being
caught, and is currently reporting Actual wire +1

In our case, it appears that the default value of 12 is too low and
calculates own position further from the stern resulting in the
incorrect wire and as such, need to adjust the wire correction to a more
suitable value.

During testing on 2022-04-23 (event:
http://132virtualwing.org/index.php/page/event?id=1764) we determined
that a value of 15 reported the correct wire for all participants during
multiple traps across all available wires and bolters