chg: fox script: Player name in message, and audible sound to hit unit
Sound is only heard by the hit unit, not all members of the group.

An updated message now includes the player's name, and is sent to the
whole group, so that flight leads etc. are aware of their WM going down
etc.

See: https://discord.com/channels/268095055724675082/608192300409159680/1046803365318906007

NB: Moose doesn't allow us to disable the notifications with the silly
death message etc. so I have edited it out of the moose library itself
and implemented the message in our event hook.

This way, if we just update moose, the worst case is we get the old
message with no information, and the new message, and know to edit it
without loosing any intended behaviours.